fibo even valued

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -1,13 +1,19 @@
+/**
+ * main - to find and prints the sum of the even-valued terms,
+ * followed by a new line
+ *
+ * Return: 0 
+ */
 #include <stdio.h>
 
-long long int evenFibSum(long long int limit) {
+long int evenFibSum(long int limit) {
   if (limit < 2) {
     return 0;
   }
 
-  long long int ef1 = 0, ef2 = 2, sum = ef1 + ef2;
+  long int ef1 = 0, ef2 = 2, sum = ef1 + ef2;
   while (ef2 <= limit) {
-    long long int ef3 = 4 * ef2 + ef1;
+    long int ef3 = 4 * ef2 + ef1;
     if (ef3 > limit) {
       break;
     }
@@ -21,8 +27,8 @@ long long int evenFibSum(long long int limit) {
 }
 
 int main() {
-  long long int limit = 4000000;
-  long long int sum = evenFibSum(limit);
-  printf("The sum of the even-valued Fibonacci terms less than or equal to %lld is %lld.\n", limit, sum);
+  long int limit = 4000000;
+  long int sum = evenFibSum(limit);
+  printf("The sum of the even-valued Fibonacci terms less than or equal to %ld is %ld.\n", limit, sum);
   return 0;
 }
